module3-task1: OOP: Create general class Component & inherit Task and TaskEdit from main Component
