+++ src/cards.js
@@ -1,5 +1,5 @@
-import {cardsContainer, createCards} from './constants';
+// import {cardsContainer, createCards} from './constants';
+//
+// const INITIAL_AMOUNT_OF_CARDS = 7;
 
-const INITIAL_AMOUNT_OF_CARDS = 7;
-
-createCards(INITIAL_AMOUNT_OF_CARDS, cardsContainer);
+// createCards(INITIAL_AMOUNT_OF_CARDS, cardsContainer);
