+++ src/index.js
@@ -1,2 +1,3 @@
 import './filters';
 import './cards';
+import './main';
